remove render targets from the Register component into their corresponding entry points
