+++ examples/basic-form/src/components/TextBox/primitive.js
@@ -1,13 +1,17 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-primitives';
 import styles from './style';
+import StrengthMeter from '../StrengthMeter';
 
-
-const TextBox = ({ label, value, children }: Props) => (
+const TextBox = ({ label, value, type }: Props) => (
   <View style={styles.formElement}>
     <Text style={styles.label}>{label}</Text>
     <View style={styles.textbox}>{value}</View>
-    {children}
+    {type === "password" && value.length > 0 && 
+      <StrengthMeter
+        password={value}
+      />
+    }
   </View>
 );
